+++ front/src/pages/plant-page/PlantPage.tsx
@@ -1,7 +1,7 @@
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import baseTheme from '../../theme.ts';
 // import "./PlantPage.css"
-import PlantCard from '../../components/Posts/PlantCard.tsx';
+import PlantCard from '../../components/Posts/PostPlantCard.tsx';
 import Grid from '@mui/material/Grid';
 import PlantReview from '../../components/PlantReview.tsx';
 import { useEffect, useState } from 'react';
